requirements.txt
awscli==1.16.202
botocore==1.12.192
colorama==0.3.9
docutils==0.14
icalendar==4.0.3
jmespath==0.9.4
pathspec==0.5.9
pyasn1==0.4.5
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
rsa==3.4.2
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
yamllint==1.16.0
